Introduce a ClassgraphWorkerModule
